ingestion/requirements.txt
garminconnect==0.2.22
supabase==2.9.0
python-dotenv==1.0.1
pytest==8.3.2
